+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 1. `doc.go` package overview documentation.
 2. Reworked logging to use uhppoted-lib/log package.
 
+### Updated
+1. Fixed static-check lint errors and warnings.
+
 
 ## [0.8.3](https://github.com/uhppoted/uhppoted-rest/releases/tag/v0.8.3) - 2022-12-16